Rename b64data to base64data
